refactor : project 조회 서비스 로직 분리
